+++ pkg/aikit2llb/inference/image.go
@@ -42,6 +42,7 @@ func emptyImage(c *config.InferenceConfig) *specs.Image {
 		"NVIDIA_DRIVER_CAPABILITIES=compute,utility",
 		"NVIDIA_VISIBLE_DEVICES=all",
 		"LD_LIBRARY_PATH=/usr/local/cuda/lib64",
+		"BUILD_TYPE=cublas",
 	}
 	if c.Runtime == utils.RuntimeNVIDIA {
 		img.Config.Env = append(img.Config.Env, cudaEnv...)
@@ -51,14 +52,13 @@ func emptyImage(c *config.InferenceConfig) *specs.Image {
 		switch c.Backends[b] {
 		case utils.BackendExllama, utils.BackendExllamaV2:
 			exllamaEnv := []string{
-				"BUILD_TYPE=cublas",
-				"EXTERNAL_GRPC_BACKENDS=exllama:/tmp/localai/backend/python/exllama/backend.py,exllama2:/tmp/localai/backend/python/exllama2/run.sh",
+				"EXTERNAL_GRPC_BACKENDS=exllama:/tmp/localai/backend/python/exllama/run.sh,exllama2:/tmp/localai/backend/python/exllama2/run.sh",
 				"CUDA_HOME=/usr/local/cuda",
 			}
 			img.Config.Env = append(img.Config.Env, exllamaEnv...)
 		case utils.BackendMamba:
 			mambaEnv := []string{
-				"EXTERNAL_GRPC_BACKENDS=mamba:/tmp/localai/backend/python/mamba/backend.py",
+				"EXTERNAL_GRPC_BACKENDS=mamba:/tmp/localai/backend/python/mamba/run.sh",
 				"CUDA_HOME=/usr/local/cuda",
 			}
 			img.Config.Env = append(img.Config.Env, mambaEnv...)
